:lipstick: Always crop the guide content
Textual 2.x broke how OptionList handles wrapping or not wrapping the
content of an OptionList. There used to be a `wrap` parameter but the
maintainer decided to remove that in 2.x.x, saying instead that you should:

  use CSS `text-wrap: nowrap; text-overflow: ellipsis`

only, this doesn't work in many cases. If you have custom Option classes
that build things out of rich Text (which would be very normal for a lot of
applications that have non-trivial OptionList widgets) it just doesn't work.

Here I handle that in a different way.