:sparkles: Add the ability to view an entry in the classic style
Closes #1 in what I think is a better way than I first intended. I had
envisaged a full Textual theme to handle this, to theme the whole
application this way; however, Textual's themes don't really lend themselves
that well to this sort of thing. Thinking on it more I think it makes sense
to have this be a novel mode for the entry viewer widget itself.