:lipstick: Remove the guide directory keys from the footer
One thing I don't like about Textual's Footer widget is that it makes its
own decisions about the order of keys in the footer. This is fine in some
situations but often you want function keys to mostly come first, but some
last too. There's no sensible way of doing this right now.

So I find the best approach is to keep as much as possible out of the
Footer. With the change I've just made to the textual-enhanced help screen I
can clean up the footer when in the guide help.