:construction: WIP: Guide-wide and directory-wide search (redux)
More work on getting the global search up and running. This mostly involves
more UI work and also code to actually walk through the guides and their
entries.

This was a worthwhile task too as I managed to uncover a couple of bugs (or
at least things that need taking into account) in ngdb.py:

  https://github.com/davep/ngdb.py/issues/17
  https://github.com/davep/ngdb.py/issues/18